[FEATURE] 뚜벗 이름 조회 및 졸업 정보 조회 api 설계

See merge request s11-final/S11P31E101!36

+++ server/userttubeot/src/main/java/com/user/userttubeot/ttubeot/domain/dto/TtubeotLogRequestDTO.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Builder
 public class TtubeotLogRequestDTO {
 
-    private Integer userTtubeotOwnershipId; // 유저의 뚜벗 아이디
+    private Long userTtubeotOwnershipId; // 유저의 뚜벗 아이디
     private Integer ttubeotLogType; // 로그타입
 
 }
